requirements.txt
streamlit
pypdf
langchain==0.1.20
langchain-openai
langchain-community
openai
faiss-cpu
python-dotenv
tiktoken